requirements.txt
fastapi  
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
pypdf2
python-dotenv
email-validator
langchain
langchain-text-splitters
langchain-community       # HuggingFaceEmbeddings wrapper
chromadb                  # vector database
sentence-transformers       # the actual bge-small model
httpx
pymupdf                          # PDF text extraction (fitz)
langchain                        # text splitter
